"ads2_m7_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2 assignments/m7/1. MST/LazyPrimMST.java
@@ -5,7 +5,7 @@ public class LazyPrimMST {
     /**
      * {total weight of MST}.
      */
-    private float weight;
+    private double weight;
     /**
      * {edges in the MST}.
      */
